Develop

+++ src/pages/SignIn/KakaoRedirect.tsx
@@ -20,17 +20,19 @@ function KakaoRedirect() {
   const signType = useSetRecoilState(loginType);
 
   useEffect(() => {
-    localStorage.clear();
-    const token = params.token || ''; // 'undefined'를 빈 문자열로 대체
-    const decoded = jwtDecode<Token>(token);
-    const parsedBody: DecodedToken = JSON.parse(decoded.body);
+    const token = params.token; // 'undefined'를 빈 문자열로 대체
+    if (token) {
+      localStorage.clear();
+      const decoded = jwtDecode<Token>(token);
+      const parsedBody: DecodedToken = JSON.parse(decoded.body);
 
-    localStorage.setItem('token', token);
-    localStorage.setItem('nickname', parsedBody.nickName);
-    localStorage.setItem('memberId', String(parsedBody.id));
+      localStorage.setItem('token', token);
+      localStorage.setItem('nickname', parsedBody.nickName);
+      localStorage.setItem('memberId', String(parsedBody.id));
 
-    signType(false);
-    window.location.replace('/main');
+      signType(false);
+      window.location.replace('/main');
+    }
   }, []);
 
   return <></>;
